+++ protected/humhub/widgets/TopMenu.php
@@ -1,29 +1,19 @@
 <?php
 
 /**
- * HumHub
- * Copyright © 2014 The HumHub Project
- *
- * The texts of the GNU Affero General Public License with an additional
- * permission and of our proprietary license can be found at and
- * in the LICENSE file you have received along with this program.
- *
- * According to our dual licensing model, this program can be used either
- * under the terms of the GNU Affero General Public License, version 3,
- * or under a proprietary license.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
  */
 
 namespace humhub\widgets;
 
+use Yii;
+use humhub\modules\user\components\User;
+
 /**
  * TopMenuWidget is the primary top navigation class extended from MenuWidget.
  *
- * @package humhub.widgets
  * @since 0.5
  * @author Luke
  */
@@ -31,11 +21,28 @@ class TopMenu extends BaseMenu
 {
 
     /**
-     * @var String template to use
+     * @inheritdoc
      */
     public $template = "topNavigation";
-    
+
+    /**
+     * @inheritdoc
+     */
     public $id = 'top-menu-nav';
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        parent::init();
+
+        // Don't show top menu if guest access is disabled
+        if (Yii::$app->user->isGuest && !User::isGuestAccessEnabled()) {
+            $this->template = '';
+        }
+    }
+
 }
 
 ?>