+++ protected/humhub/widgets/TopMenuRightStack.php
@@ -1,33 +1,36 @@
 <?php
 
 /**
- * HumHub
- * Copyright © 2014 The HumHub Project
- *
- * The texts of the GNU Affero General Public License with an additional
- * permission and of our proprietary license can be found at and
- * in the LICENSE file you have received along with this program.
- *
- * According to our dual licensing model, this program can be used either
- * under the terms of the GNU Affero General Public License, version 3,
- * or under a proprietary license.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
  */
 
 namespace humhub\widgets;
 
+use Yii;
+use humhub\modules\user\components\User;
+
 /**
  * TopMenuRightStackWidget holds items like search (right part)
  *
- * @package humhub.widgets
  * @since 0.6
  * @author Luke
  */
 class TopMenuRightStack extends BaseStack
 {
+
+    /**
+     * @inheritdoc
+     */
+    public function run()
+    {
+        // Don't show stack if guest access is disabled and user is not logged in
+        if (Yii::$app->user->isGuest && !User::isGuestAccessEnabled()) {
+            return;
+        }
+
+        return parent::run();
+    }
 
 }